Implementing server client setup
Added flask server, and client application to access the server
