+++ Server/requirements.txt
@@ -1,2 +1,3 @@
 requests==2.28.1
-Flask==2.0.2
+Flask==2.2.2
+Werkzeug==2.3.7
