+++ query.py
@@ -1,6 +1,7 @@
 import requests
 
-def query(address, key):
-    header = {"x-api-key" : key}
-    response = requests.get(address, headers=header)
+def query(address):
+    #header = {"x-api-key" : key}
+    #response = requests.get(address, headers=header)
+    response = requests.get(address)
     return response.json()
